Updating code in one swoop because I'm lazy

+++ src/delete_files.py
@@ -6,6 +6,7 @@ import os
 import sys
 import metadata
 import argparse
+import unicodedata
 
 settings = None
 
@@ -13,6 +14,7 @@ def get_settings():
   ''' Return command-line settings '''
   parser = argparse.ArgumentParser(description='Delete files in corpus not listed in metadata')
   parser.add_argument('-i', dest='input', required=True, help='Input CSV of metadata describing files')
+  parser.add_argument('-t', '--test', dest='test', action='store_true', help='Do a test run; only list files to be deleted.')
   return parser.parse_args()
 
 def delete_files(md):
@@ -21,17 +23,29 @@ def delete_files(md):
   for row in md:
     directories.append(os.path.dirname(row['filename']))
     all_files.append(row['filename'])
-  directories = set(directories)
+  directories = frozenset(directories)
+  all_files = frozenset(all_files)
 
   for basedir in directories:
     for root, subdirs, files in os.walk(basedir):
       for filename in files:
-        path = os.path.join(root, filename)
-        if filename not in all_files:
-          os.remove(path)
+        # Note: Mac OS X stores files as fully decomposed unicode
+        # whereas our metadata strings are 'normal form composed'
+        # @see: http://stackoverflow.com/questions/16467479/normalizing-unicode
+        # @see: http://apple.stackexchange.com/questions/10476/how-to-enter-special-characters-so-that-bash-terminal-understands-them
+        path = unicodedata.normalize('NFC', os.path.join(root, filename))
+        if path not in all_files:
+          if settings.test:
+            print(path)
+          else:
+            print('Removing {}'.format(path))
+            os.remove(path)
     if not os.listdir(root):
       # remove empty directories
-      os.rmdir(root)
+      if settings.test:
+        print(root)
+      else:
+        os.rmdir(root)
 
 def main():
   global settings
